backend/foodgram/requirements.txt
Django==3.2.16
requests==2.18.4
djangorestframework==3.12.4
djoser==2.1.0
django-filter==22.1
pytz==2022.1
drf-extra-fields==3.4.1
isort==5.11.4
Pillow==8.3.2
webcolors==1.13
reportlab==3.6.1
django-split-settings==1.2.0
python-dotenv==1.0.0
python-decouple==3.8
uuid==1.30
urllib3==1.22
idna==2.6
gunicorn==20.1.0
psycopg2-binary==2.8.6
django-cors-headers===4.0.0